Add AWS API test for multi deletion
